Updates Layout and Image_Browser classes
