+++ Layouts.py
@@ -1,9 +1,10 @@
 from PyQt5.QtWidgets import QHBoxLayout, QGridLayout
+from Image import Image
 
-class Thumbnail_Layout(QHBoxLayout):
+class Thumbnail_layout(QHBoxLayout):
     def __init__(self):
-        super().__init__()
+        super.__init__()
 
-class Zoomed_Layout(QGridLayout):
+class Zoomed_layout(QGridLayout):
     def __init__(self):
-        super().__init__()
+        super.__init__()
